move test to node directory
