--- tests/sdk/java/testdata/await/generic/yaml/Pulumi.yaml
@@ -1,60 +1,0 @@
-name: generic-await-logic
-runtime: yaml
-description: YAML version of our generic await stack for a faster local dev loop.
-resources:
-  ns:
-    type: kubernetes:core/v1:Namespace
-    properties:
-      metadata:
-        name: generic-await
-
-  provider:
-    type: pulumi:providers:kubernetes
-    properties:
-      namespace: ${ns.metadata.name}
-
-  crds:
-    type: kubernetes:yaml/v2:ConfigFile
-    options:
-      provider: ${provider}
-    properties:
-      file: ../crds.yaml
-
-  resources:
-    # ConfigGroup is a workaround for https://github.com/pulumi/pulumi-kubernetes/issues/1971
-    type: kubernetes:yaml/v2:ConfigGroup
-    options:
-      provider: ${provider}
-      dependsOn:
-        - ${crds}
-    properties:
-      yaml: |
-        apiVersion: test.pulumi.com/v1
-        kind: GenericAwaiter
-        metadata:
-          name: wants-ready-condition
-          annotations:
-            pulumi.com/patchForce: "true"
-            pulumi.com/timeoutSeconds: "20"
-        spec:
-          someField: untouched
-        status:
-          conditions:
-            - type: Ready
-              status: "False"
-
-        apiVersion: test.pulumi.com/v1
-        kind: GenericAwaiter
-        metadata:
-          name: wants-generation-increment
-          generation: 3
-          annotations:
-            pulumi.com/patchForce: "true"
-            pulumi.com/timeoutSeconds: "20"
-        spec:
-          someField: untouched
-        status:
-          conditions:
-            - type: Ready
-              status: "True"
-          observedGeneration: 2
