+++ tests/sdk/nodejs/testdata/await/generic/Pulumi.yaml
