+++ tests/sdk/nodejs/testdata/await/generic/crds.yaml
