+++ tests/sdk/nodejs/testdata/await/generic/make-ready.sh
@@ -2,4 +2,4 @@
 set -ex
 
 kubectl patch --type=merge -n=generic-await "genericawaiters.test.pulumi.com" wants-ready-condition -p '{"status": {"conditions": [{"type": "Ready", "status": "True"}]}}'
-kubectl patch --type=merge -n=generic-await "genericawaiters.test.pulumi.com" wants-generation-increment -p '{"status": {"observedGeneration": 3}}'
+kubectl patch --type=merge -n=generic-await "genericawaiters.test.pulumi.com" wants-generation-increment -p '{"status": {"observedGeneration": 6}}'
