+++ tests/sdk/nodejs/testdata/await/generic/package.json
